adminProduct_Management_4
Changes to the frontend (html, css, javascript) of the view, search, edit, and add pages. A Side bar was added to allow for easier data flow and changes were also made to the functionality of the form page such as replacing the input tags of Type and Supplier with the select tag.  Text in the view products page was made larger so its clearer to read.